Created new TaskReviewController in new framework
Move listChallenges into it.
