+++ app/org/maproulette/controllers/api/TaskReviewController.scala
@@ -8,7 +8,7 @@ import javax.inject.Inject
 import org.maproulette.Config
 import org.maproulette.data.ActionManager
 import org.maproulette.exception.NotFoundException
-import org.maproulette.framework.model.{Challenge, ChallengeListing, User, Project}
+import org.maproulette.framework.model.{Challenge, User, Project}
 import org.maproulette.framework.service.ServiceManager
 import org.maproulette.framework.psql.Paging
 import org.maproulette.models.Task
@@ -55,8 +55,6 @@ class TaskReviewController @Inject() (
       changeService,
       bodyParsers
     ) {
-
-  implicit val challengeListingWrites: Writes[ChallengeListing] = Json.writes[ChallengeListing]
 
   /**
     * Gets and claims a task that needs to be reviewed.
@@ -427,78 +425,4 @@ class TaskReviewController @Inject() (
       }
     }
   }
-
-  /**
-    * Returns a list of challenges that have reviews/review requests.
-    *
-    * @param reviewTasksType  The type of reviews (1: To Be Reviewed,  2: User's reviewed Tasks, 3: All reviewed by users)
-    * @param tStatus The task statuses to include
-    * @param excludeOtherReviewers Whether tasks completed by other reviewers should be included
-    * @return JSON challenge list
-    */
-  def listChallenges(
-      reviewTasksType: Int,
-      tStatus: String,
-      excludeOtherReviewers: Boolean = false,
-      limit: Int,
-      page: Int
-  ): Action[AnyContent] =
-    Action.async { implicit request =>
-      this.sessionManager.authenticatedRequest { implicit user =>
-        val taskStatus = tStatus match {
-          case v if v.nonEmpty => Utils.toIntList(v)
-          case _               => None
-        }
-
-        val challenges = this.serviceManager.challengeListing.withReviewList(
-          reviewTasksType,
-          user,
-          taskStatus,
-          excludeOtherReviewers,
-          Paging(limit, page)
-        )
-
-        // Populate some parent/virtual parent project data
-        val projects = Some(
-          this.serviceManager.project
-            .list(challenges.map(c => c.parent))
-            .map(p => p.id -> p)
-            .toMap
-        )
-
-        var vpIds = scala.collection.mutable.Set[Long]()
-        challenges.map(c => {
-          c.virtualParents match {
-            case Some(vps) =>
-              vps.map(vp => vpIds += vp)
-            case _ => // do nothing
-          }
-        })
-        val vpObjects =
-          this.serviceManager.project.list(vpIds.toList).map(p => p.id -> p).toMap
-
-        val jsonList = challenges.map { c =>
-          val projectJson = Json
-            .toJson(projects.get(c.parent))
-            .as[JsObject] - Project.KEY_GROUPS
-
-          var updated =
-            Utils.insertIntoJson(Json.toJson(c), Challenge.KEY_PARENT, projectJson, true)
-          c.virtualParents match {
-            case Some(vps) =>
-              val vpJson =
-                Some(
-                  vps.map(vp => Json.toJson(vpObjects.get(vp)).as[JsObject] - Project.KEY_GROUPS)
-                )
-              updated = Utils.insertIntoJson(updated, Challenge.KEY_VIRTUAL_PARENTS, vpJson, true)
-            case _ => // do nothing
-          }
-          updated
-        }
-
-        Ok(
-          Json.toJson(jsonList)
-        )
-      }
-    }
 }
